Refactor mission order backend
