+++ worlds/sc2/mission_order/layout_types.py
@@ -4,7 +4,7 @@ import math
 from abc import ABC, abstractmethod
 
 if TYPE_CHECKING:
-    from .structs import SC2MOGenMission
+    from .nodes import SC2MOGenMission
 
 class LayoutType(ABC):
     size: int
